Compiles cleanly, seems to run all right, but some of the pixels do not seem to work correctly.

+++ src/image_funcs.h
@@ -3,6 +3,6 @@
 
 void image_to_ppm(Grapher *grapher, char *filename);
 void image_to_raw(Grapher *grapher, char *filename);
-void choose_RGB(double pixel, double max_pixel, double *rgb);
+void choose_RGB(double pixel, double max_pixel, int *rgb);
 
 #endif /* IMAGE_H_ */
